feat: parking management service with enhanced webhook handling and database operations
- Introduced `PersistenceHandler` for centralized database operation handling with error logging.
- Updated `ParkingSpotRepositoryAdapter` to utilize `PersistenceHandler` for finding parking spots.
- Changed `ParkingEntity` and `ParkingEventEntity` to use `BigDecimal` for monetary values.
- Created `ParkingCustomQueryRepository` and its implementation for custom database queries related to parking.
- Refactored `ParkingEventRepository` to include new query methods for event retrieval.
- Implemented new webhook handlers (`EntryWebhookHandler`, `ParkedWebhookHandler`, `ExitWebhookHandler`) for processing parking events.
- Consolidated webhook handling logic into `ParkingWebhookHandler`.
- Added `CalculatePricingMultiplierHandler` for dynamic pricing calculations based on occupancy.
- Removed the old `WebhookService` and replaced it with a more modular approach using specific handlers.
- Updated database schema to accommodate new fields for pricing in `parking_events`.
- Enhanced exception handling with new messages for better user feedback.

+++ src/main/kotlin/io/github/martinsjavacode/parkingmanagement/domain/gateway/client/response/ParkingDataResponse.kt
@@ -2,12 +2,13 @@ package io.github.martinsjavacode.parkingmanagement.domain.gateway.client.respon
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect
 import com.fasterxml.jackson.annotation.JsonProperty
+import java.math.BigDecimal
 
 @JsonAutoDetect
 data class ParkingDataResponse(
     val sector: String,
     @JsonProperty("base_price")
-    val basePrice: Double,
+    val basePrice: BigDecimal,
     @JsonProperty("max_capacity")
     val maxCapacity: Int,
     @JsonProperty("open_hour")
