+++ src/main/kotlin/io/github/martinsjavacode/parkingmanagement/domain/model/Parking.kt
@@ -1,12 +1,13 @@
 package io.github.martinsjavacode.parkingmanagement.domain.model
 
 import kotlinx.coroutines.flow.Flow
+import java.math.BigDecimal
 import java.time.LocalTime
 
 data class Parking(
     val id: Long?,
     val sector: String,
-    val basePrice: Double,
+    val basePrice: BigDecimal,
     val maxCapacity: Int,
     val openHour: LocalTime,
     val closeHour: LocalTime,
